+++ src/my/mymoviesamn/JInternalFrame_Statistics.java
@@ -219,7 +219,7 @@ public class JInternalFrame_Statistics extends javax.swing.JInternalFrame {
             for (Movie movie:movies){
             //Βρίσκω από την κάθε Αγαπημένη Λίστα την Ταινία με την υψηλότερη Βαθμολογία
             if (i== movie.getFavoriteListId().getId()){
-                String rating = Double.toString(movie.getRating());
+                String rating = Float.toString(movie.getRating());
                 model.addRow(new String[]{movie.getTitle(),rating});
                 break;
                 }
@@ -246,8 +246,8 @@ public class JInternalFrame_Statistics extends javax.swing.JInternalFrame {
         model.setColumnIdentifiers(new String[]{"Ταινία","Βαθμολογία"});
         //Με ένα for διαβάζω τη Λίστα και την περνάω στον βοηθητικό πίνακα
         movies.forEach((movie) -> {
-            //αφού μετατρέψω τη βαθμολογία από double σε string
-            String rating = Double.toString(movie.getRating());
+            //αφού μετατρέψω τη βαθμολογία από float σε string
+            String rating = Float.toString(movie.getRating());
             //Προσθέτω μία μία γραμμή της Λίστας στον Πίνακα
             model.addRow(new String[]{movie.getTitle(),rating});
         });
